+++ Tests/SwiftLintFrameworkTests/Resources/FileHeaderRuleFixtures/FileNameMatchingComplex.swift
@@ -1,0 +1,3 @@
+// Copyright © 2016
+// File: "FileNameMatchingComplex.swift"
+struct A {}