+++ Tests/SwiftLintFrameworkTests/Resources/FileHeaderRuleFixtures/FileNameMatchingSimple.swift
@@ -1,0 +1,3 @@
+// FileNameMatchingSimple.swift
+// Copyright © 2016
+struct A {}